+++ pyserum/market/__init__.py
@@ -1,3 +1,5 @@
+import pyserum.market.instructions
+from .enums import OrderType, Side # noqa: F401
 from .market import Market  # noqa: F401
 from .orderbook import OrderBook  # noqa: F401
 from .state import MarketState as State  # noqa: F401
