+++ pyserum/market/open_orders_account.py
@@ -10,8 +10,8 @@ from solana.rpc.types import Commitment, MemcmpOpts
 from solana.system_program import CreateAccountParams, create_account
 from solana.transaction import TransactionInstruction
 
-from .._layouts.open_orders import OPEN_ORDERS_LAYOUT
-from ..instructions import DEFAULT_DEX_PROGRAM_ID
+from .._layouts.market.open_orders import OPEN_ORDERS_LAYOUT
+from .instructions import DEFAULT_DEX_PROGRAM_ID
 from ..utils import load_bytes_data
